requirements.txt
streamlit
pandas
Pillow
PyGithub
